Invert RedactEventJSON
