+++ eventcrypto_test.go
@@ -41,7 +41,7 @@ func TestVerifyEventSignatureTestVectors(t *testing.T) {
 	}
 
 	testVerifyOK := func(input string) {
-		redactedInput, err := RedactEventJSON([]byte(input), RoomVersionV1)
+		redactedInput, err := RoomVersionV1.RedactEventJSON([]byte(input))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -52,7 +52,7 @@ func TestVerifyEventSignatureTestVectors(t *testing.T) {
 	}
 
 	testVerifyNotOK := func(reason, input string) {
-		redactedInput, err := RedactEventJSON([]byte(input), RoomVersionV1)
+		redactedInput, err := RoomVersionV1.RedactEventJSON([]byte(input))
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -390,7 +390,7 @@ func TestVerifyAllEventSignatures(t *testing.T) {
 	if len(verifier.requests) != 1 {
 		t.Fatalf("Number of requests: got %d, want 1", len(verifier.requests))
 	}
-	wantContent, err := RedactEventJSON(eventJSON, RoomVersionV1)
+	wantContent, err := RoomVersionV1.RedactEventJSON(eventJSON)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -448,7 +448,7 @@ func TestVerifyAllEventSignaturesForInvite(t *testing.T) {
 	if len(verifier.requests) != 2 {
 		t.Fatalf("Number of requests: got %d, want 2", len(verifier.requests))
 	}
-	wantContent, err := RedactEventJSON(eventJSON, RoomVersionV1)
+	wantContent, err := RoomVersionV1.RedactEventJSON(eventJSON)
 	if err != nil {
 		t.Fatal(err)
 	}
